reqrite audience component
