+++ src/components/shared/AudiencePicker/AudienceDropdownSelector/styles.module.scss
@@ -1,0 +1,28 @@
+.multiValue {
+    display: flex;
+    align-items: center;
+    // background-color: var(--events-options-background-color);
+
+    .icon {
+        color: white;
+        align-self: stretch;
+        padding: 0 3px;
+        padding-right: 2px;
+        line-height: 0;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        &.departmentType {
+            background-color: var(--ifm-color-info);
+        }
+        &.classType {
+            background-color: var(--ifm-color-warning);
+        }
+        &.classGroup {
+            background-color: var(--ifm-color-blue);
+        }
+        &.unknwon {
+            background-color: var(--ifm-color-danger);
+        }
+    }
+}
