--- src/components/shared/AudiencePicker/Department/styles.module.scss
@@ -1,9 +1,0 @@
-.departmentClasses {
-    .department {
-        display: flex;
-        flex-wrap: wrap;
-    }
-    .year {
-        display: flex;
-    }
-}
