+++ src/components/shared/AudiencePicker/index.tsx
@@ -8,12 +8,12 @@ import { observer } from 'mobx-react-lite';
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
 import { default as DepartmentModel } from '@site/src/models/Department';
-import Department from './Department';
+import Department from './DepartmentTab/Department';
 import Button from '../Button';
 import _ from 'lodash';
 import Translate, { translate } from '@docusaurus/Translate';
-import Info from '../../Event/EventFields/AudienceOptions/Info';
-import AudienceSelector from './AuienceDropdownSelector';
+import AudienceSelector from './AudienceDropdownSelector';
+import DepartmentTab from './DepartmentTab';
 
 interface Props {
     event: EventModel;
@@ -21,6 +21,7 @@ interface Props {
 }
 
 const AudiencePicker = observer((props: Props) => {
+    const { event } = props;
     const departmentStore = useStore('departmentStore');
     const userStore = useStore('userStore');
     const { current } = userStore;
@@ -28,8 +29,6 @@ const AudiencePicker = observer((props: Props) => {
         return null;
     }
 
-    const departments = departmentStore.groupedByLetter;
-    const { event } = props;
     const error = event.errorFor('audience');
     const {
         someDepartments,
@@ -51,7 +50,6 @@ const AudiencePicker = observer((props: Props) => {
                         Schulen/Klassen
                     </Translate>
                 </h4>
-                <AudienceSelector event={event} />
                 <div className={clsx(styles.flex)}>
                     <Button
                         text={translate({
@@ -94,29 +92,12 @@ const AudiencePicker = observer((props: Props) => {
                         }}
                     />
                 </div>
-                <Tabs className={clsx(styles.tabs)} lazy>
-                    {Object.keys(departments)
-                        .sort()
-                        .map((letter, idx) => {
-                            const color = (departments[letter] as DepartmentModel[])[0].color;
-                            const touched = (departments[letter] as DepartmentModel[]).some((d) =>
-                                d.classes.some((c) => event.affectsClass(c))
-                            );
-                            return (
-                                <TabItem
-                                    value={letter}
-                                    label={departmentStore.letterToName(letter)}
-                                    key={letter}
-                                    attributes={{
-                                        className: clsx(touched && styles.touched),
-                                        style: { color: color }
-                                    }}
-                                >
-                                    <Department departments={departments[letter]} event={event} />
-                                </TabItem>
-                            );
-                        })}
-                </Tabs>
+                <DepartmentTab
+                    departments={departmentStore.groupedByLetter}
+                    event={event}
+                    className={clsx(styles.tabItems)}
+                />
+                <AudienceSelector event={event} />
                 {error && <div className={clsx(styles.errorMessage)}>{error.message}</div>}
             </div>
         </div>
