+++ src/components/shared/Button/styles.module.scss
@@ -8,6 +8,9 @@
     border-color: var(--ifm-button-border-color);
     background: var(--ifm-button-background-color);
     color: var(--ifm-button-color);
+    &.noBorder {
+        border-width: 0;
+    }
     &:disabled {
         pointer-events: auto;
     }
