+++ src/components/shared/Tooltip/index.tsx
@@ -13,7 +13,7 @@ interface Props {
 }
 
 const MOUSE_ENTER_DELAY_MS = 250;
-const MOUSE_LEAVE_DELAY_MS = 200;
+const MOUSE_LEAVE_DELAY_MS = 0;
 
 const Tooltip = observer((props: Props) => {
     if (!props.title || props.disabled) {
@@ -27,7 +27,7 @@ const Tooltip = observer((props: Props) => {
             on="hover"
             position={['top center', 'top right', 'top left']}
             arrow={false}
-            offsetY={5}
+            offsetY={8}
             contentStyle={{ width: 'max-content', maxWidth: '50%' }}
         >
             <div className={clsx(styles.tooltip)}>{props.title}</div>
