+++ src/models/Event.ts
@@ -25,7 +25,8 @@ import {
     getLastMonday,
     DAYS_LONG,
     dateBetween,
-    formatDateLong
+    formatDateLong,
+    currentGradeYear
 } from './helpers/time';
 import Klass from './Untis/Klass';
 import Lesson from './Untis/Lesson';
@@ -459,7 +460,7 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
 
     @action
     toggleClass(klass: KlassName) {
-        this.setClass(klass, !this.affectedClassNames.has(klass));
+        this.setClass(klass, !this.classes.has(klass));
     }
 
     @action
@@ -473,36 +474,11 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
         } else if (currentActive && !value) {
             this.classes.delete(klass);
         }
-        if (!value) {
-            const classGroupName = klass.slice(0, 3);
-            if (this.classGroups.has(classGroupName)) {
-                this.classGroups.delete(classGroupName);
-                const allOfGroup = this.store.root.untisStore.findClassesByGroupName(classGroupName);
-                allOfGroup.forEach((c) => {
-                    if (c.name !== klass) {
-                        this.classes.add(c.name);
-                    }
-                });
-            }
-            const department = this.departments.find((d) => d.classes.some((c) => c.name === klass));
-            if (department) {
-                this.departmentIds.delete(department.id);
-                department.classes.forEach((c) => {
-                    if (c.name !== klass) {
-                        this.setClass(c.name, true);
-                    }
-                });
-            }
-        }
-
-        this.normalizeAffectedClasses();
     }
 
     @action
     toggleClassGroup(groupName: string) {
-        const current = this.untisStore.classesGroupedByGroupNames.get(groupName);
-        const isActive =
-            this.classGroups.has(groupName) || current?.every((c) => this.affectedClassNames.has(c.name));
+        const isActive = this.classGroups.has(groupName);
         this.setClassGroup(groupName, !isActive);
     }
 
@@ -518,21 +494,6 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
         } else if (currentActive && !value) {
             this.classGroups.delete(klassGroup);
         }
-
-        if (!value) {
-            const affectedDeps = this.departments.filter((d) => d.classGroups.has(klassGroup));
-            if (affectedDeps.length > 0) {
-                affectedDeps.forEach((dep) => {
-                    this.departmentIds.delete(dep.id);
-                    dep.classes.forEach((c) => {
-                        if (!c.name.startsWith(klassGroup)) {
-                            this.classes.add(c.name);
-                        }
-                    });
-                });
-            }
-        }
-        this.normalizeAffectedClasses();
     }
 
     /**
@@ -583,44 +544,17 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
     }
 
     @action
-    normalizeAffectedClasses() {
-        const cNames = new Set<string>(this.affectedClassNames);
-        const classes = new Set<KlassName>();
-        const classGroups = new Set<string>();
-        const departmentIds = new Set<string>();
-        this.departmentStore.departments.forEach((d) => {
-            if (d.classes.length > 1 && d.classes.every((c) => cNames.has(c.name))) {
-                d.classes.forEach((c) => cNames.delete(c.name));
-                departmentIds.add(d.id);
-            }
-        });
-        this.untisStore.classesGroupedByGroupNames.forEach((classes, group) => {
-            if (classes.length > 1 && classes.every((c) => cNames.has(c.name))) {
-                classes.forEach((c) => cNames.delete(c.name));
-                classGroups.add(group);
-            }
-        });
-        this._unknownClassGroups.forEach((cg) => {
-            classGroups.add(cg);
-            cNames.forEach((c) => {
-                if (c.startsWith(cg)) {
-                    cNames.delete(c);
-                }
-            });
-        });
-        cNames.forEach((c) => {
-            if (c.length === 4) {
-                classes.add(c as KlassName);
-            }
-        });
-        this.classes.replace(classes);
-        this.classGroups.replace(classGroups);
-        this.departmentIds.replace(departmentIds);
+    toggleDepartment(department: Department) {
+        this.setDepartment(department, !this.departmentIds.has(department.id));
     }
 
     @action
-    toggleDepartment(department: Department) {
-        this.setDepartment(department, !this.departmentIds.has(department.id));
+    setDepartmentId(departmentId: string, setActive: boolean) {
+        const dep = this.store.root.departmentStore.find<Department>(departmentId);
+        if (!dep) {
+            return;
+        }
+        this.setDepartment(dep, setActive);
     }
 
     @action
@@ -634,7 +568,6 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
         } else if (!currentActive && setActive) {
             this.departmentIds.add(department.id);
         }
-        this.normalizeAffectedClasses();
     }
 
     /**
@@ -652,6 +585,39 @@ export default class Event extends ApiModel<EventProps, ApiAction> implements iE
     get affectedDepartments() {
         const depIds = new Set([...this.departmentIds, ...this.untisClasses.map((c) => c.departmentId)]);
         return this.store.getDepartments([...depIds]);
+    }
+
+    @computed
+    get currentGradeYear() {
+        return currentGradeYear(this.start);
+    }
+
+    /**
+     * describes how many grade years should be additionally affected by this event.
+     */
+    @computed
+    get gradeYearRange() {
+        return currentGradeYear(this.end) - this.currentGradeYear;
+    }
+
+    @computed
+    get affectedClassGroups() {
+        const classGroups = new Set<string>(this.classGroups);
+        this.departments.forEach((d) => {
+            const isActive = this.store
+                .getDepartmentsByLetter(d.letter)
+                .map((d) => d.id)
+                .every((id) => this.departmentIds.has(id));
+            if (!isActive) {
+                return;
+            }
+            for (let i = 0; i < d.schoolYears; i++) {
+                const year = `${this.currentGradeYear + i}`.slice(2);
+                const group = `${year}${d.letter}`;
+                classGroups.add(group);
+            }
+        });
+        return classGroups;
     }
 
     /**
